OboeTester: refactor automated tests
Add AutomatedTestRunner and use it for Test Disconnect
and Auto Glitch tests.
Add DataPaths stub.
