+++ apps/OboeTester/app/src/main/AndroidManifest.xml
@@ -98,6 +98,12 @@
             android:screenOrientation="portrait">
         </activity>
 
+        <activity
+            android:name="com.google.sample.oboe.manualtest.TestDataPathsActivity"
+            android:label="@string/title_data_paths"
+            android:screenOrientation="portrait">
+        </activity>
+
         <service
             android:name="com.google.sample.oboe.manualtest.AudioMidiTester"
             android:permission="android.permission.BIND_MIDI_DEVICE_SERVICE">
