+++ apps/OboeTester/app/src/main/res/layout/activity_data_paths.xml
@@ -8,30 +8,17 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.google.sample.oboe.manualtest.AutoGlitchActivity">
+    tools:context="com.google.sample.oboe.manualtest.TestDisconnectActivity">
 
-    <LinearLayout
+    <TextView
+        android:id="@+id/text_instructions"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="horizontal">
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/duration"
-            />
-        <Spinner
-            android:id="@+id/spinner_glitch_duration"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:entries="@array/glitch_times"
-            />
-        <TextView
-            android:id="@+id/actualNativeApi"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/seconds_per_test"
-            />
-    </LinearLayout>
+        android:lines="1"
+        android:text="@string/test_disconnect_instructions"
+        android:textColor="#F44336"
+        android:textSize="18sp"
+        android:textStyle="bold" />
 
     <com.google.sample.oboe.manualtest.AutomatedTestRunner
         android:id="@+id/auto_test_runner"
